Playing around with dyn Callable and Any
This commit implements Value::Callable as using a trait dyn Callable, and downcasts on specific cases using Any. This is useful, and a general replacement of Value as a Box<dyn Value> trait should be considered.

There are some caveats as well, for example the downcast_ref() implementation on dyn Callable.