Playing around with proc-macro
Considering about a proc_macro_attribute for taking [tokay_method] or similar annotations as base for Tokay methods. Maybe this could be useful later when a method signature might be directly generated from the method definition. But after a while of testing around, I think that a build-script written in Tokay is a better way to generate Tokay builtins directly from the code.