Tests for using Tokay v0.4 in proc-macro macros
This commit uses Tokay v0.4 itself to parse Tokay-like call definitions inside the proc-macro to identify buitlins. This is part of Tokay 0.5.

The main reason for this change was, that the previous attempt didn't use Rust syntax inside the macro, so that `cargo fmt` ignores them.

This isn't a wanted behavior. Maybe this will be fixed in Tokay 0.6 or befre, I'm not sure right now.