New Op::Sep VM instruction
Op::Sep ensures that the TOS value is a separate value and not a variable, so that `i = 0; ++i ++i ++i` correctly generates (1, 2, 3) and not (3, 3, 3), as the last value of i was shared by all stack items.

The compiler inserts Op::Sep only in sitations where the result of a sequence must ensured to be separate.