Op::Extend and Op::Capture
This makes all inline-sequences consuming input to areas with an in-place reader extend, which avoid the outer frame to break because input is not indicated to be parsed.

Nevertheless, programs like

```tokay
(Int ',' Int) | Ident   # will accept: 1,2a
```
will accept inputs like `1,2a`, which is obviously not a wanted behavior.