Phew. okay. lets call this an intermediate commit.
Well... this thing was - and still is - nasty.

This state does allow to call any callable from a builtin.
And implements a mapped iterator by the way.

Here, this works:

	for i in range(2, 100).map(@x { print(x) x*x }) print(i)

Anyway, I'm not sure if this is the best way.

iter_map() generates another iter, which is borrowed when the provided function is executed. This leads into the problem, that in case map makes use of the same iterator, program will crash. This isn't useful.