Generate any `StoreGlobal*`-ops correctly
This enforces that any global variable is handled using global VM statements all the time. In case of an undeclared global, it was initially e.g. stored as local. This is no problem with the current implementation, but might become a problem in future.

This commit is a hack with the loop, first tried another way but this was too much refactoring and didn't simplify anything. Hopefully later.