refactor: Scopes and intermediate object construction in Compiler (#132)
* Undo debug string evaluation

* Replace enum by struct Scope

This uses the same interface than before, but removes the unpacking overhead.

* Move usages into scope

* Intermediate commit

1. Move variables and tempos into ImlParseletModel.
2. Scratch Scope2 as next scoping implementation, already compiles.
3. Unsure if this is all right and good, but its better than before.

* Playing around with the Scope2 draft

* wip: New compiler scoping design

This is almost working and the conceptual way to go, after testing several other ways and stuff.

[skip ci]

* Adding Scope compatibility layer to Compiler

Makes Tokay useable again! Nevertheless, this is only a crutch to restore the previous behavior. Further considerations are necessary to bring Compiler and Scope together optimally.